requirements.txt
google-genai>=1.55.0
pillow 
numpy